+++ lib/go-tc/server_server_capability.go
@@ -27,6 +27,12 @@ type ServerServerCapability struct {
 	ServerCapability *string    `json:"serverCapability" db:"server_capability"`
 }
 
+// MultipleServerCapabilities represents an association between a server and list of server capabilities.
+type MultipleServerCapabilities struct {
+	ServerID           int      `json:"serverId" db:"server"`
+	ServerCapabilities []string `json:"serverCapabilities" db:"server_capability"`
+}
+
 // ServerServerCapabilitiesResponse is the type of a response from Traffic
 // Ops to a request made to its /server_server_capabilities.
 type ServerServerCapabilitiesResponse struct {